+++ tests/test_audioframe.py
@@ -1,5 +1,6 @@
 import warnings
 
+import numpy
 from av import AudioFrame
 from av.deprecation import AttributeRenamedWarning
 
@@ -79,6 +80,7 @@ class TestAudioFrameConveniences(TestCase):
     def test_basic_to_ndarray(self):
         frame = AudioFrame(format='s16p', layout='stereo', samples=160)
         array = frame.to_ndarray()
+        self.assertEqual(array.dtype, '<i2')
         self.assertEqual(array.shape, (2, 160))
 
     def test_basic_to_nd_array(self):
@@ -93,3 +95,88 @@ class TestAudioFrameConveniences(TestCase):
         self.assertEqual(
             str(recorded[0].message),
             'AudioFrame.to_nd_array is deprecated; please use AudioFrame.to_ndarray.')
+
+    def test_ndarray_dbl(self):
+        layouts = [
+            ('dbl', 'mono', '<f8', (1, 160)),
+            ('dbl', 'stereo', '<f8', (1, 320)),
+            ('dblp', 'mono', '<f8', (1, 160)),
+            ('dblp', 'stereo', '<f8', (2, 160)),
+        ]
+        for format, layout, dtype, size in layouts:
+            array = numpy.ndarray(shape=size, dtype=dtype)
+            for i in range(size[0]):
+                array[i][:] = numpy.random.rand(size[1])
+            frame = AudioFrame.from_ndarray(array, format=format, layout=layout)
+            self.assertEqual(frame.format.name, format)
+            self.assertEqual(frame.layout.name, layout)
+            self.assertEqual(frame.samples, 160)
+            self.assertTrue((frame.to_ndarray() == array).all())
+
+    def test_ndarray_flt(self):
+        layouts = [
+            ('flt', 'mono', '<f4', (1, 160)),
+            ('flt', 'stereo', '<f4', (1, 320)),
+            ('fltp', 'mono', '<f4', (1, 160)),
+            ('fltp', 'stereo', '<f4', (2, 160)),
+        ]
+        for format, layout, dtype, size in layouts:
+            array = numpy.ndarray(shape=size, dtype=dtype)
+            for i in range(size[0]):
+                array[i][:] = numpy.random.rand(size[1])
+            frame = AudioFrame.from_ndarray(array, format=format, layout=layout)
+            self.assertEqual(frame.format.name, format)
+            self.assertEqual(frame.layout.name, layout)
+            self.assertEqual(frame.samples, 160)
+            self.assertTrue((frame.to_ndarray() == array).all())
+
+    def test_ndarray_s16(self):
+        layouts = [
+            ('s16', 'mono', '<i2', (1, 160)),
+            ('s16', 'stereo', '<i2', (1, 320)),
+            ('s16p', 'mono', '<i2', (1, 160)),
+            ('s16p', 'stereo', '<i2', (2, 160)),
+        ]
+        for format, layout, dtype, size in layouts:
+            array = numpy.random.randint(0, 256, size=size, dtype=dtype)
+            frame = AudioFrame.from_ndarray(array, format=format, layout=layout)
+            self.assertEqual(frame.format.name, format)
+            self.assertEqual(frame.layout.name, layout)
+            self.assertEqual(frame.samples, 160)
+            self.assertTrue((frame.to_ndarray() == array).all())
+
+    def test_ndarray_s16p_align_8(self):
+        frame = AudioFrame(format='s16p', layout='stereo', samples=159, align=8)
+        array = frame.to_ndarray()
+        self.assertEqual(array.dtype, '<i2')
+        self.assertEqual(array.shape, (2, 159))
+
+    def test_ndarray_s32(self):
+        layouts = [
+            ('s32', 'mono', '<i4', (1, 160)),
+            ('s32', 'stereo', '<i4', (1, 320)),
+            ('s32p', 'mono', '<i4', (1, 160)),
+            ('s32p', 'stereo', '<i4', (2, 160)),
+        ]
+        for format, layout, dtype, size in layouts:
+            array = numpy.random.randint(0, 256, size=size, dtype=dtype)
+            frame = AudioFrame.from_ndarray(array, format=format, layout=layout)
+            self.assertEqual(frame.format.name, format)
+            self.assertEqual(frame.layout.name, layout)
+            self.assertEqual(frame.samples, 160)
+            self.assertTrue((frame.to_ndarray() == array).all())
+
+    def test_ndarray_u8(self):
+        layouts = [
+            ('u8', 'mono', 'u1', (1, 160)),
+            ('u8', 'stereo', 'u1', (1, 320)),
+            ('u8p', 'mono', 'u1', (1, 160)),
+            ('u8p', 'stereo', 'u1', (2, 160)),
+        ]
+        for format, layout, dtype, size in layouts:
+            array = numpy.random.randint(0, 256, size=size, dtype=dtype)
+            frame = AudioFrame.from_ndarray(array, format=format, layout=layout)
+            self.assertEqual(frame.format.name, format)
+            self.assertEqual(frame.layout.name, layout)
+            self.assertEqual(frame.samples, 160)
+            self.assertTrue((frame.to_ndarray() == array).all())